逐步成形

+++ src/middleware/asd.go
@@ -27,7 +27,7 @@ var (
 func CheckLogin(logger log.Logger) endpoint.Middleware {
 	return func(next endpoint.Endpoint) endpoint.Endpoint {
 		return func(ctx context.Context, request interface{}) (response interface{}, err error) {
-			ctx = context.WithValue(ctx, UserIdContext, 1)
+			ctx = context.WithValue(ctx, UserIdContext, int64(1))
 			return next(ctx, request)
 		}
 	}
